+++ docs/release_notes.rst
@@ -8,6 +8,12 @@ Read up on all the latest improvements.
 eth-utils v5.3.0 (2025-04-14)
 -----------------------------
 
+Breaking Changes
+~~~~~~~~~~~~~~~~
+
+- Removes ``is_list`` and ``is_tuple`` from the types module.
+
+
 Bugfixes
 ~~~~~~~~